requirements.txt
# APIs e autenticação
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
O365==2.0.26
python-dotenv==1.0.0

# Manipulação de calendários
icalendar==5.0.7
recurring-ical-events==2.0.2
pytz==2023.3

# Utilitários
pydantic==2.4.2
loguru==0.7.0